mcp_servers/kanban/backend-v2/pyproject.toml
[project]
name = "nova-kanban-mcp"
version = "0.1.0"
description = "Nova Kanban MCP Server with PostgreSQL backend"
authors = [
    {name = "Nova Team", email = "daniel.kuehlwein@gmail.com"},
]
dependencies = [
    "fastmcp>=2.3.4",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.23",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dateutil>=2.8.2",
    "uuid6>=2023.5.2",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.13"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "pytest-postgresql>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 